+++ Assets/Scripts/MB/ConstructionReservation.cs
@@ -3,5 +3,5 @@ using System.Collections;
 
 public class ConstructionReservation : Reservation {
     public string resourceTag;
-    public float amount;
+    public decimal amount;
 }
