+++ Assets/Scripts/Orders/Simple/ReserveStorageOrder.cs
@@ -16,7 +16,7 @@ public class ReserveStorageOrder : BaseOrder {
 
     public override void DoStep() {
         string type = resource.typeTag;
-        float amount = resource.amount;
+        decimal amount = resource.amount;
 
         if (actor.gameController.ReserveStorage(actor, type, amount)) {
             Debug.Log("Reserved storage");
